Extract native agent lifecycle into native_agents module.
Unify dual/selective compile, scan .mars/agents once per sync, and drop
unnecessary lock/outcome clones in compile() so compiler/mod.rs stays small.
